Add validation when datatable structure changed
