+++ Editor/DataDriven/DataTableRowView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e3d783ebb73d6c4caa2ba5f4f7b69e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
