+++ Editor/Serialization/SerializedObjectWrapperManager.cs
@@ -2,30 +2,30 @@ using System;
 using UnityEngine;
 namespace Kurisu.Framework.Serialization.Editor
 {
-    [Serializable]
-    public class GenericSerializedObjectWrapper<T> : SerializedObjectWrapper
-    {
-        [SerializeField]
-        T m_Value;
-
-        public override object Value
-        {
-            get { return m_Value; }
-            set { m_Value = (T)value; }
-        }
-    }
-
     /// <summary>
-    /// Api for editor serialization
+    /// Class to manage SerializedObjectWrapper
     /// </summary>
-    public static class SerializationEditorManager
+    public static class SerializedObjectWrapperManager
     {
+        [Serializable]
+        public class SerializedObjectWrapper<T> : SerializedObjectWrapper
+        {
+            [SerializeField]
+            T m_Value;
+
+            public override object Value
+            {
+                get { return m_Value; }
+                set { m_Value = (T)value; }
+            }
+        }
         public static SerializedObjectWrapper CreateWrapper(Type type, ref SoftObjectHandle softObjectHandle)
         {
             if (type == null) return null;
 
             var wrapper = GlobalObjectManager.GetObject(softObjectHandle) as SerializedObjectWrapper;
-            if (!wrapper)
+            // Validate wrapped type
+            if (!wrapper || wrapper.Value.GetType() != type)
             {
                 wrapper = Wrap(Activator.CreateInstance(type));
                 GlobalObjectManager.UnregisterObject(softObjectHandle);
@@ -37,14 +37,21 @@ namespace Kurisu.Framework.Serialization.Editor
         {
             GlobalObjectManager.UnregisterObject(softObjectHandle);
         }
-        public static SerializedObjectWrapper GetWrapper(SoftObjectHandle softObjectHandle)
+        public static SerializedObjectWrapper GetWrapper(Type type, SoftObjectHandle softObjectHandle)
         {
-            return GlobalObjectManager.GetObject(softObjectHandle) as SerializedObjectWrapper;
+            var wrapper = GlobalObjectManager.GetObject(softObjectHandle) as SerializedObjectWrapper;
+            // Validate wrapped type
+            if (wrapper && wrapper.Value.GetType() != type)
+            {
+                GlobalObjectManager.UnregisterObject(softObjectHandle);
+                return null;
+            }
+            return wrapper;
         }
         private static SerializedObjectWrapper Wrap(object value = null)
         {
             Type type = value.GetType();
-            Type genericType = typeof(GenericSerializedObjectWrapper<>).MakeGenericType(type);
+            Type genericType = typeof(SerializedObjectWrapper<>).MakeGenericType(type);
             Type dynamicType = DynamicTypeBuilder.MakeDerivedType(genericType, type);
             var dynamicTypeInstance = ScriptableObject.CreateInstance(dynamicType);
             if (dynamicTypeInstance is not SerializedObjectWrapper wrapper)
